- use R.bool.show_sura_names_translation as default value instead of hardcoded. - move the showing translated name from the adapter to the fragment.
