+++ app/src/main/java/com/quran/labs/androidquran/ui/fragment/SuraListFragment.java
@@ -8,6 +8,7 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.DefaultItemAnimator;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -28,6 +29,7 @@ import javax.inject.Inject;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.Disposable;
 import io.reactivex.observers.DisposableSingleObserver;
+
 import org.jetbrains.annotations.NotNull;
 
 import static com.quran.labs.androidquran.data.Constants.JUZ2_COUNT;
@@ -46,6 +48,8 @@ public class SuraListFragment extends Fragment {
 
   @Inject QuranInfo quranInfo;
   @Inject QuranDisplayData quranDisplayData;
+  @Inject QuranSettings quranSettings;
+  boolean showSuraTranslatedName = false;
   private int numberOfPages;
 
   public static SuraListFragment newInstance() {
@@ -88,8 +92,11 @@ public class SuraListFragment extends Fragment {
   public void onResume() {
     final Activity activity = getActivity();
     if (activity instanceof QuranActivity) {
-      QuranSettings settings = QuranSettings.getInstance(activity);
-      ((QuranListAdapter) mRecyclerView.getAdapter()).showHideSuraTranslatedName(settings.isShowSuraTranslatedName());
+      boolean newValueOfShowSuraTranslatedName = quranSettings.isShowSuraTranslatedName();
+      if (this.showSuraTranslatedName != newValueOfShowSuraTranslatedName) {
+        showHideSuraTranslatedName(newValueOfShowSuraTranslatedName);
+        this.showSuraTranslatedName = newValueOfShowSuraTranslatedName;
+      }
       disposable = ((QuranActivity) activity).getLatestPageObservable()
           .first(Constants.NO_PAGE)
           .observeOn(AndroidSchedulers.mainThread())
@@ -109,7 +116,7 @@ public class SuraListFragment extends Fragment {
             }
           });
 
-      if (settings.isArabicNames()) {
+      if (quranSettings.isArabicNames()) {
         updateScrollBarPositionHoneycomb();
       }
     }
@@ -129,8 +136,7 @@ public class SuraListFragment extends Fragment {
 
     Activity activity = getActivity();
     boolean wantPrefix = activity.getResources().getBoolean(R.bool.show_surat_prefix);
-    QuranSettings settings = QuranSettings.getInstance(activity);
-    boolean wantTranslation = settings.isShowSuraTranslatedName();
+    boolean wantTranslation = quranSettings.isShowSuraTranslatedName();
     for (int juz = 1; juz <= JUZ2_COUNT; juz++) {
       final String headerTitle = activity.getString(R.string.juz2_description,
           QuranUtils.getLocalizedNumber(activity, juz));
@@ -154,5 +160,33 @@ public class SuraListFragment extends Fragment {
     }
 
     return elements;
+  }
+
+  public void showHideSuraTranslatedName(boolean showSuraTranslatedName) {
+    QuranRow[] elements = ((QuranListAdapter) mRecyclerView.getAdapter()).getElements();
+    for (QuranRow row : elements) {
+      if (row.sura != 0) {
+        if (showSuraTranslatedName) {
+          addTranslatedNameToSuraName(row);
+        } else {
+          removeSuraTranslatedName(row);
+        }
+      }
+    }
+    mRecyclerView.getAdapter().notifyDataSetChanged();
+  }
+
+  private void removeSuraTranslatedName(QuranRow row) {
+    if (row.text.indexOf("(") != -1)
+      row.text = row.text.substring(0, row.text.indexOf("(") - 1);
+  }
+
+  private void addTranslatedNameToSuraName(QuranRow row) {
+    StringBuilder builder = new StringBuilder(row.text);
+    String translation = getResources().getStringArray(R.array.sura_names_translation)[row.sura - 1];
+    if (!TextUtils.isEmpty(translation)) {
+      builder.append(" (" + translation + ")");
+    }
+    row.text = builder.toString();
   }
 }
